Overhaul + new import system

+++ src/ecs/ComponentPool.h
@@ -7,32 +7,32 @@ public:
 	ComponentPool(const u32 compSize);
 
 	template<typename Component>
-	Component& Get(const u32 index);
+	Component& Get(const size_t index);
 
 	template<typename Component>
-	Component& Add(const u32 index);
+	Component& Add(const size_t index);
 private:
-	void* GetComponentAddress(const u32 index) const;
+	void* GetComponentAddress(const size_t index) const;
 private:
 	u32 m_CompSize;
 	u32 m_NextCompIndex;
 	
-	u32 m_CurBufferSize;
+	size_t m_CurBufferSize;
 	Ref<u8[]> m_Buffer;
 
-	u32 m_ComponentOffsetsSize;
+	size_t m_ComponentOffsetsSize;
 	Ref<u32[]> m_ComponentOffsets;
 };
 
 template <typename Component>
-Component& ComponentPool::Get(const u32 index) {
+Component& ComponentPool::Get(const size_t index) {
 	return *static_cast<Component*>(GetComponentAddress(index));
 }
 
 template <typename Component>
-Component& ComponentPool::Add(const u32 index) {
+Component& ComponentPool::Add(const size_t index) {
 	if (index >= m_ComponentOffsetsSize) {
-		const u32 oldSize = m_ComponentOffsetsSize;
+		const size_t oldSize = m_ComponentOffsetsSize;
 		m_ComponentOffsetsSize = index + 1;
 		
 		const Ref<u32[]> tempComponentOffsets = m_ComponentOffsets;
@@ -41,10 +41,10 @@ Component& ComponentPool::Add(const u32 index) {
 	}
 
 	if (m_NextCompIndex >= m_CurBufferSize) {
-		const u32 oldSize = m_CurBufferSize;
+		const size_t oldSize = m_CurBufferSize;
 		
 		constexpr f32 resizeScaler = 1.5;
-		m_CurBufferSize = std::ceil(m_CurBufferSize * resizeScaler);
+		m_CurBufferSize = static_cast<size_t>(std::ceil(m_CurBufferSize * resizeScaler));
 
 		const Ref<u8[]> tempBuffer = m_Buffer;
 		m_Buffer = MakeRef<u8[]>(m_CompSize * m_CurBufferSize);
